fix(vpcgw):better vizualisation gw network dhcp
