+++ internal/namespaces/vpcgw/v1/custom.go
@@ -11,7 +11,8 @@ func GetCommands() *core.Commands {
 
 	human.RegisterMarshalerFunc(vpcgw.GatewayNetworkStatus(""), human.EnumMarshalFunc(gatewayNetworkStatusMarshalSpecs))
 	human.RegisterMarshalerFunc(vpcgw.GatewayStatus(""), human.EnumMarshalFunc(gatewayStatusMarshalSpecs))
-	human.RegisterMarshalerFunc(vpcgw.Gateway{}, gatewayNetworkMarshallerFunc)
+	human.RegisterMarshalerFunc(vpcgw.Gateway{}, gatewayMarshalerFunc)
+	human.RegisterMarshalerFunc(vpcgw.GatewayNetwork{}, gatewayNetworkMarshalerFunc)
 
 	cmds.MustFind("vpc-gw", "gateway-type", "list").Override(vpcgwGatewayTypeListBuilder)
 	cmds.MustFind("vpc-gw", "gateway", "create").Override(gatewayCreateBuilder)
