+++ internal/namespaces/vpcgw/v1/custom_gateway.go
@@ -42,3 +42,23 @@ func gatewayCreateBuilder(c *core.Command) *core.Command {
 	}
 	return c
 }
+
+func gatewayMarshalerFunc(i interface{}, opt *human.MarshalOpt) (string, error) {
+	type tmp vpcgw.Gateway
+	vpcgtw := tmp(i.(vpcgw.Gateway))
+	opt.Sections = []*human.MarshalSection{
+		{
+			FieldName: "IP",
+			Title:     "IP",
+		},
+		{
+			FieldName: "GatewayNetworks",
+			Title:     "GatewayNetworks",
+		},
+	}
+	str, err := human.Marshal(vpcgtw, opt)
+	if err != nil {
+		return "", err
+	}
+	return str, nil
+}
